+++ lib/urchin/context.ex
@@ -34,6 +34,7 @@ defmodule Urchin.Context do
     assigns: %{},
     min_log_level: "debug",
     expose_internal_errors: false,
+    validate_arguments: false,
     cancelled_ref: nil
   ]
 
@@ -52,6 +53,7 @@ defmodule Urchin.Context do
           assigns: map(),
           min_log_level: String.t(),
           expose_internal_errors: boolean(),
+          validate_arguments: boolean(),
           cancelled_ref: reference() | nil
         }